pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plsec"
version = "0.1.0"
description = "Security tooling for AI coding assistants"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Peerlabs", email = "security@peerlabs.io" }
]
keywords = [
    "security",
    "ai",
    "claude",
    "opencode",
    "llm",
    "coding-assistant",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
plsec = "plsec.cli:app"

[project.urls]
Homepage = "https://github.com/peerlabs/plsec"
Documentation = "https://github.com/peerlabs/plsec#readme"
Repository = "https://github.com/peerlabs/plsec"
Issues = "https://github.com/peerlabs/plsec/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/plsec"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "S",    # flake8-bandit
]
ignore = [
    "S101",  # assert used (fine in tests)
    "S603",  # subprocess without shell=True (we check this)
    "S607",  # partial path (we use shutil.which)
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=plsec --cov-report=term-missing"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]

[tool.uv.sources]
# Add any local or git sources here if needed
